Remove Deprecated Client and Server Headers
- Deleted `MCP_Client.h`, `MCP_Server.h`, and `Server__DT.h` files to streamline the codebase and eliminate unused components.
- This cleanup aligns with recent refactoring efforts to modernize the protocol and session interfaces, enhancing overall maintainability.
- Updated related files to reflect the removal of these headers, ensuring consistency across the codebase.
