+++ Source/Public/Sandbox/ClientBase.h
@@ -10,11 +10,12 @@ MCP_NAMESPACE_BEGIN
 class IClientAPI {
   public:
     // Tool Operations
-    virtual MCPTask<vector<Tool>> ListTools(const optional<string>& InCursor = {}) = 0;
+    virtual MCPTask<ListToolsResult> ListTools(const optional<ListToolsRequest>& Params = nullopt);
     virtual MCPTask<ToolResult> CallTool(const ToolCall& InCall) = 0;
 
     // Resource Operations
-    virtual MCPTask<vector<Resource>> ListResources(const optional<string>& InCursor = {}) = 0;
+    virtual MCPTask<ListResourcesResult>
+    ListResources(const optional<ListResourcesRequest>& Params = nullopt);
     virtual MCPTask<vector<ResourceContent>> ReadResource(const string& InURI) = 0;
     virtual MCPTask<vector<ResourceTemplate>>
     ListResourceTemplates(const optional<string>& InCursor = {}) = 0;
@@ -24,9 +25,10 @@ class IClientAPI {
     virtual MCPTask_Void UnsubscribeFromResource(const string& InURI) = 0;
 
     // Prompt Operations
-    virtual MCPTask<vector<Prompt>> ListPrompts(const optional<string>& InCursor = {}) = 0;
-    virtual MCPTask<PromptMessage>
-    GetPrompt(const string& InName, const unordered_map<string, string>& InArguments = {}) = 0;
+    virtual MCPTask<ListPromptsResult>
+    ListPrompts(const optional<ListPromptsRequest>& Params = nullopt);
+    virtual MCPTask<GetPromptResult>
+    GetPrompt(const GetPromptRequest::GetPromptRequestParams& Params) = 0;
 
     // Sampling (for servers to request LLM operations from clients)
     virtual MCPTask<SamplingResult> CreateMessage(const SamplingRequest& InRequest) = 0;
@@ -36,8 +38,9 @@ class IClientAPI {
         function<MCPTask<SamplingResult>(const SamplingRequest&)> InHandler) = 0;
 
     // Root Directory Management
-    virtual MCPTask<vector<string>> ListRoots() = 0;
-    virtual void SetRoots(const vector<string>& InRoots) = 0;
+    virtual MCPTask<ListRootsResult> ListRoots() = 0;
+    virtual MCPTask_Void SetRoots(const vector<string>& InRoots) = 0;
+    virtual MCPTask_Void SendRootsListChanged() = 0;
 
     // Notification Handlers
     virtual void OnToolListChanged(function<void()> InCallback) = 0;
@@ -47,14 +50,59 @@ class IClientAPI {
     virtual void OnRootsListChanged(function<void()> InCallback) = 0;
 
     // Capabilities
-    virtual void SetCapabilities(const Capability& InCapabilities) = 0;
-    virtual Capability GetCapabilities() const = 0;
+    virtual void SetCapabilities(const ClientCapabilities& InCapabilities) = 0;
+    virtual ClientCapabilities GetCapabilities() const = 0;
+
+    // TODO: Determine if below is right or above is right. This was used in Direct Translation
+    // Registers new capabilities. This can only be called before connecting to a transport.
+    //
+    // The new capabilities will be merged with any existing capabilities previously given (e.g., at
+    // initialization).
+    void RegisterCapabilities(const ClientCapabilities& Capabilities);
 };
 
+/**
+ * An MCP client on top of a pluggable transport.
+ *
+ * The client will automatically begin the initialization flow with the server when Connect() is
+ * called.
+ *
+ * To use with custom types, extend the base Request/Notification/Result types and pass them as type
+ * parameters.
+ */
 class ClientBase : public IClientAPI, public IMCP {
   public:
+    struct ClientOptions : public ProtocolOptions {
+        optional<ClientCapabilities> Capabilities;
+    };
+    // Initializes this client with the given name and version information.
+    ClientBase(const Implementation& ClientInfo, const optional<ClientOptions>& Options = nullopt);
+
     ClientBase(const ClientOptions& InOptions);
     ~ClientBase() override;
+
+    // Client Methods
+    // TODO: Ping?
+    // TODO: MCPTask<CompleteResult> Complete(const CompleteRequest& Params);
+    // TODO: MCPTask_Void SetLoggingLevel(LoggingLevelRequest& InRequest);
+    MCPTask_Void Connect(optional<shared_ptr<Transport>> InTransport);
+
+    // TODO: This was used in Direct Translation - Needed?
+  protected:
+    void AssertCapability(const string& Capability, const string& Method);
+    void AssertCapabilityForMethod(const string& Method);
+    void AssertNotificationCapability(const string& Method);
+    void AssertRequestHandlerCapability(const string& Method);
+
+    // TODO: This was used in Direct Translation - Needed?
+    // private:
+    //   void CacheToolOutputSchemas(const vector<Tool>& Tools);
+    //   optional<ValidateFunction> GetToolOutputValidator(const string& ToolName);
+
+  private:
+    Implementation m_ClientInfo;
+    ClientCapabilities m_Capabilities;
+    // unordered_map<string, ValidateFunction> m_CachedToolOutputValidators; TODO: Needed?
 };
 
 MCP_NAMESPACE_END
